Let's give bun a try
Bun is a new JavaScript runtime with everything built in.
It seems like it could be a lot easier than using a bunch of tools.
It also is a bit faster and has some improvements to security.
This replaces node, c8, esbuild, taze, maybe some other things.

This commit also upgrades our eslint config to work with v9.